Account for tenants with no schema during pruning
In the event an RBAC service pod restarts during the creation of a schema, there is
an edge case exposed where a `Tenant` object will be created, but not a schema.

We cannot have both the `Tenant` and schema creation in a transaction currently,
because of a deadlock issue which this PR helped to resolve: https://github.com/RedHatInsights/insights-rbac/pull/519

A more robust fix would be to know when a schema has not been created for a given
tenant, and remove the `Tenant` object. This is a stop-gap until we move off
of schema-based tenancy in the near future, to allow us to manually remove these
via Turnpike.

It ensures that in the check to see if a tenant has been modified, we confirm
the schema exists. We need to do this because if we don't, and it doesn't exist,
the `search_path` on the query will fall back to using the public schema, in
which case there will be custom groups/roles, and we won't be able to propertly
prune the tenant which has been created without a valid schema.