Return custom default group for tenant when serving from public
Currently when we get `groups_for_principal`, we query `Group.platform_default_set()`
which will return all `platform_default=True` groups, which is what we need when
we serve from the tenant-specific schemas. This is because we modify the default
group per schema.

When we serve from the public schema, we need to check to see if there's a custom
default group in the public schema for that tenant. If so, we'll use that, otherwise
return the default group from the public schema.

We need to supply the `request` object to the `utils` methods, which this updates
the callers of as well.