Add internal endpoing for initializing tenants
This will add `POST /_private/api/tenant/<schema-name>/init/` which will be an
idempodent way to do the following for a tenant:

- create a schema and run migrations if one does not exists for the tenant
- run seeds after the schema is configured
- flip `ready=True`

This will allow us to manage `Tenant` objects which are in a "hung" state, by
running the schema creation and migrations/seeds if the schema does not exist.
If it does exist, it will skip the schema creation and migration, and just run
seeds, which will not be an issue.

Tenants are able to get into this state when the service restarts during the tenant
creation process. We're not able to run the tenant object creation, and schema
setup in a transaction because of an edge-case with concurrent requests during
creation. This means we have the potential to orphan Tenant records, leaving them
in a state where `ready=False`, and subsequent requests will hang.