Ensure we include system (public) roles when service from public in the role queryset
Prior to this change, when serving from public, the role queryset scopes _all_ role
to the tenant's roles. This means the role queryset will _never_ include any of
the public/system roles.

This change ensures that when we serve from public, we get all roles in the base
queryset from the public schema and the tenant.