Fix the access cache key when serving from public
Since we always set the access policy cache key based on the `request.tenant.schema_name` [1]
which is what we want so that it's specific to each account, we need to ensure that
when we invalidate the cache, we're using the correct key.

These changes will ensure we key off the schema name from the object in the signal,
vs the case where the `public` schema is being used and the current code will
cause a miss on the key lookup.

[1] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/management/access/view.py#L108