Always run init container
This will ensure that now we always run the init container when deploying with
Clowder, regardless of environment. The init container will run seeds and migrations.
Currently we still need to explicitly target the public schema, but once DTS
is removed and the updates from #660 are in, we will be able to update our commands
to just run Django migrations with the default command, and seeds will not run
against all schemas.

- updates the Makefile to only run migrations again public (local only)
- updates README for consistency and to avoid confusion
- in the init container script, no longer check for ephemeral env, and run seeds/migrationss
  in the context of the public schema
- remove `EPH_ENV` from the template
- update the Openshift run script (even though we're not currently using this)
- update the command args to accept a list of `schema_list`

We'll still want to:
- remove the `EPH_ENV` params from app-interface deployments