Handle no payload in `/roles/<uuid>/` PATCH
Currently we allow an empty payload in `/roles/<uuid>/` `PATCH` requests. However
if you supply _no_ payload, the `json.loads(request.data)` call fails with a 500
because it cannot convert an empty string.

In order to preserve existing behavior, and align an empty payload to behave like
no payload (a no-op) in `PATCH` requests, we'll default the payload to en empty
object.