Add internal endpoint for querying/deleting invalid default admin groups
Surfaces the following internal endpoint in Turnpike:
```
/_private/api/utils/invalid_default_admin_groups/
```
which can be requested with a `GET` or `DELETE` method. A `GET` will return all
invalid groups' names as well as a count. A `DELETE` (when enabled) will remove
the groups.

This is to remediate data related to https://github.com/RedHatInsights/insights-rbac/pull/694