Fix `/roles/<uuid>/access/` endpoint
The `/roles/<uuid>/access/` endpoint now returns a 404 for system roles.

This appears to be happening because the roles lookup for this view is filtering
on the request tenant [1], which means in the case of system roles, a 404 will
be returned because the role UUID matches, but the tenant is the public tenant,
which will not match on the request tenant.

The UUID should be enough to grab the correct role.

[1] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/management/role/view.py#L407