Creating a "Custom default access" group should only use public tenant roles by default
Currently the query to grab roles for a custom default group [1] has an implicit
expectation that any `platform_default` roles will be coming from the public tenant
as that's the only way they can be created via seeds.

There's an edge-case where in stage, as part of our public schema data migration,
one role from a non-public tenant (an old `platform_default` role from the tenant's
schema) was copied into the public schema, still tied to that tenant.

This in turn means that when a new custom default group is cut in stage, it will
include that duplicate group since the filter doesn't explicitly require the
tenant be the public tenant.

These changes close that edge-case.

To test:
- on master, manually create a duplicate `platform_default` role, which already exists
  for the public tenant, and update the `tenant_id` to be that of another tenant
- update the "Default access" group, which will create a "Custom default access"
  group, and note that if you query `/rbac/api/v1/groups/<uuid>/roles/` you'll
  see the duplicate role from the non-public tenant
- check these changes out and run the same test as above, and note the duplicate
  role does not exist on the new "Custom default access" group

[1] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/management/group/definer.py#L100