Add `?username_only=true|false` param on `/principals/`
This adds a `?username_only=` query param on `/principals/` in order to return
only the `username` values for principals. The immediate need for this is for
OCM to bypass calling IT which would return superfluous data at a performance
cost.

A new `/principals/?username_only=true` response payload would look like:
```
{
  "meta": {
    "count": 3
  },
  "links": {
    "first": "/api/rbac/v1/principals/?limit=10&offset=0",
    "next": "/api/rbac/v1/principals/?limit=10&offset=10",
    "previous": "/api/rbac/v1/principals/?limit=10&offset=0",
    "last": null
  },
  "data": [
    {
      "username": "foo"
    },
    {
      "username": "bar"
    },
    {
      "username": "baz"
    }
  ]
}
```

The code changes here set `username_only` on the options passed into the principal
proxy after validation, in order to preserve any other filters provided in the request.

From there, the proxy will short-circuit any calls to the BoP and return the locally
stored `username` values in the payload if `username_only` is `true`.