Add `core` app to house shared application logic
It seems to be convention to create an app in Django (commonly `core`) when you
have a need to share logic across applications in your service.

This does that for a shared `utils` module. We have `utils` modules in our other
applications, but could stand to refactor some of the logic into this (and other)
shared module(s).

Initially this surfaces `destructive_ok()`, potentially for use in #746