Use the Clowder-provided Kafka ca cert when available
When `LoadedConfig.kafka.brokers[0].cacert` has a value, we should set `ssl_cafile` [1]
on the producer. We can create a tmp file with the cert in it via `app-common-python` [2]
and include that in our `KAFKA_AUTH` config.

This also updates the initialization of `KAFKA_AUTH` to be a dictionary by default,
as an empty dictionary will evaluate to `False` [3] meaning our check against
`settings.KAFKA_AUTH` [4] will continue to evaluate to `False` when we leave `KAFKA_AUTH`
initialized as an empty dictionary, which makes the logic easier to reason about.
Then we can just `.update()` the dictionary for `sasl`, and add the cert when it
exists.

[1] https://kafka-python.readthedocs.io/en/master/apidoc/KafkaProducer.html
[2] https://github.com/RedHatInsights/app-common-python/blob/ce2e7014254d53939ddcc26a84d362d511461588/README.md?plain=1#L23
[3] https://docs.python.org/2/library/stdtypes.html#truth-value-testing
[4] https://github.com/RedHatInsights/insights-rbac/blob/53ff73e4c87993fd3b6e0ab325c1fe062fa295af/rbac/core/kafka.py#L41