Add `USE_CLOWDER_CA_FOR_BOP` with default `False`
When set to `True`, this will use the internal/Clowder CA path for TLS requests
to the back-office proxy, when inside a Clowd environment.

We'll default this to false, since by default we'll need to continue to use
the existing CA for the BOP at `management/principal/certs/client.pem`, but when
configured to use MBOP or another service inside the cluster, we can flip this
to `True` and use the Clowder-provided CA path.