Add `external_source_id` to /principals/ GET
This returns `external_source_id` as part of a `/principals/` GET request, in order
to tie user records back to the external source of truth.

These IDs should be populated by bop/mbop regardless of the source of truth in
a given environment and is already part of the API spec for bop.

We may want to include a source name mapping at some point as well, but since this
is environment specific, it should be implicit knowledge from whoever is consuming
and requiring this new field.