Deduplicate access permissions list returned to clients
It is useful for clients to handle the simplest list of permissions
objects, for performance reasons.  It reduces the data transferred to
the client application, and it reduces the possibility that a client
will misinterpret some part of the data.  It also reduces the amount of
data RBAC stores in its cache, thus saving space and time.

This implements a basic process to deduplicate the list of permissions
objects returned to the client.  It does two basic optimisations:

1. Two exactly matching permissions - e.g. 'app:*:read' and 'app:*:read'
   - are combined.  Their resourceDefinitions are appended without
     attempting to interpret them.
2. The '*' verb implies all other verbs.  Therefore 'app:*:read' will be
   ignored in favour of 'app:*:*'.  Ignored permissions have their
   resourceDefinitions thrown away because they are superseded.

Further optimisations may be possible but may be app-dependent and we
should tread with caution here.

A basic test is included, but probably not yet working.

Signed-off-by: Paul Wayper <paulway@redhat.com>