POC for adding a PDP (policy decision point) in RBAC
This is currently conditionally applied to the `/access/` endpoint via a query
param, but could also be broken out into a new endpoint. This is just to illustrate
potential implementation.

To exercise the "new" functionality, hit the access endpoint with the following pattern,
aligning with some permissions your user will have locally:
```
/access/?application=approval&resource_type=workflows&verb=read&pdp=true
```

The `pdp=true` param will force the new POC path to return a payload similar to the
following, based on a composite lookup of permissions for the authenticated user,
the `application`, `resource_type` and `verb` supplied:
```
{
  "has_access": true,
  "permission": "approval:workflows:read",
  "application": "approval",
  "resource_type": "workflows",
  "verb": "read"
}
```

There are also a few things to consider this does not account for:
- optimal performance
- resourceDefinitions
- multiple permissions
- `OR` conditions in query params