feature: return stubs for the service accounts
The UI team is blocked because RBAC still doesn't fully support the
service accounts. Even though the ADR is nearly finished, the sooner we
can let the UI team consume the endpoints with the service accounts the
better.

They have told us that it is not required to return real data, so that
is why we are returning static stubs instead. Once the ADR is finished,
we will remove them.

The default behavior is to honor old query parameters, that is why they
have precedence over the rest.

RHCLOUD-28261