fix: missing query parameter to remove service accounts
The UI team has let us know that the service accounts query parameter is
missing from the OpenAPI spec, which causes, obviously, the API client
generators to not generate it.

Therefore, there is no way for them to remove the service accounts
because the query parameter is not available to them.

RHCLOUD-27878