fix: order by is wrongly specified
The order by query parameter is required to be specified as "username" or "-username"
in the view, and then gets converted to an internal "sort order" option
with the "asc" or "desc" values. However, in order to apply those to the
username field the service accounts' code needs to reconvert those "asc"
and "desc" values to the "username" and "-username" values that Django
understands.

RHCLOUD-27878