fix: return consistent service accounts when bypassing IT calls
When bypassing IT calls for service accounts, we were taking the service
accounts principals from the database, and then faking an IT call
response for those service accounts, to decorate the principal with that
mocked data. However we were generating a UUID every time for the
fields, when we could have simply used the stored UUID for the service
account, which would yield us with consistent results every time we
called the endpoint.

RHCLOUD-27878