Add initial decorator for syncing to SpiceDB (mocked)
This adds `/management/decorators.py` with a `SpiceDb` class which can organize
any decorators for SpiceDb, initially `sync`.

This decorator can be applied to any DRF view action method to ensure the data
is sent to SpiceDb (currently the generic mocked endpoint).

Within a transaction, we run the default view method, which should save the record
to the PRBAC database. If that succeeds, we'll then `POST` the view's response
data to SpiceDB, which will have the database object in JSON. From there, SpiceDB
can do what it needs to.

If the response back from SpiceDB is a non-200, we'll raise and catch a status
error. In this case we'll roll the PRBAC transaction back, and set the decorator
return value to a custom SpiceDB error response with a 424 status for "failed dependency".

If there's any other exception raised within the transaction, we'll roll the PRBAC
transaction back, and set the decorator return value to a custom generic error
response with a 424 status for "failed dependency".

What gets returned back from the decorator will be a `Response` object either from
the original view method, or the error handling, and the client will see the correct
response accordingly.

The database action in PRBAC will only be committed if and when a 200 response
from the SpiceDB service is obtained.