fix: requester's org admin condition is used for fetching default groups
When a username is specified for the query sets, we need to verify if
that username is an organization administrator in order to feed the
results with the default groups.

However, for service accounts we do not need to feed those default
results because service accounts are not considered organization
administrators.

In our check we were checking the requester's organization administrator
condition, instead of the specified username's condition.

RHCLOUD-29641