fix: unable to add service accounts to groups
Due to the recent refactoring changes which moved the token validation
to the middleware, only requests coming from service accounts were
getting their token validated and then assigned to the "User" internal
object.

This caused that the bearer token coming from the requests of any other
user would not be added to the "User" object, which in turn caused the
IT Service to start sending empty bearer tokens to IT.

Finally, IT responded with 401 Unauthorized errors because of the
missing bearer tokens.

RHCLOUD-31077