fix: stop validating the audience claim
We don't have a unified audience claim that we can validate for the
incoming tokens. Depending on which method is used to request a token,
the audience claim changes, and that causes the tokens to not be
validated correctly.

There are ongoing discussions about this specific claim and in which
way we need to validate it, but until we decide what to do, we can
safely rely on the service accounts scope that we're already validating
to make sure that the token is valid.

RHCLOUD-31103