feature: validate tokens per handler basis and per scope basis
Moving the token validation to the middleware made sense because right
now RBAC only has endpoints which require the "api.iam.service_accounts"
scope. However, that might change in the future, and we might need to
restrict the scope of the tokens depending on which view handler is
being called.

That is why we need to move back the token validation on a per handler
basis, by specifying which scopes are needed on each handler. That
way, we will be able to specify different scopes on different handlers
and the token validation will keep working.

The exception handling has been moved to the custom exception handler in
order to make our lives easier and to avoid repeating the same exception
handling over and over again.

RHCLOUD-31207