fix: service accounts not being validated
When validating a single service account, we were sending requests to IT
filtered by that single service account to avoid fetching the entire
collection. However, due to a bug, the collections were always being
returned empty.

Until the bug is solved, we need to fetch the entire collections
instead.

RHCLOUD-31265