fix: the access endpoint requires a bearer token
We were attempting to validate the service account usernames with the
bearer token from the user, but that was creating some problems with the
access endpoint.

For example, when an organization administrator was querying the access
endpoint to check the permissions of a particular service account, due
to how the "get_principal_from_request" and "get_principal" functions
work, we were requesting a bearer token for that. Another case was when the
user that generated the request was a service account.

However, in both of these cases we should be validating the service
account username not with an incoming token we were expecting, but with
a "behind the curtains" process as the one we have for the regular
principals. Because, otherwise, we are forcing the users to send us a
particular token with the "api.iam.service_accounts" scope when they're
not really trying to manage service accounts.

This, however, is not ready from the IT side so until that is provided
to us, we are going to skip the service accounts username validation in
these two functions. This is fine because the handlers that take care of
managing service accounts are still validating that a token with the
proper "api.iam.service_accounts" scope is coming with the request.

RHCLOUD-31275