Add GET list operation on `/workspaces/` for v2 API to TypeSpec file
This will generate the spec for returning a list of `BasicWorkspace` objects. It
also adds an optional query param `depth`, specifying an enum and union of any
positive int.

The enum consists of a named `All` key with a supplied value of `-1` which indicates
the entire tree will be returned. There is also a `Current` key which will return
the top-level only. The unioned `unit16` will allow for positive ints to be
supplied to drill down more granular as needed.