Bring mapping and initial role read within locked transaction to prevent out of order replication
It turns out that the initial role state was being queried outside of the atomic block and without a row lock and thus subject to a concurrent update causing policy, access, or mappings to not be the current state by the time the outbox event is written, which would create inconsistency. This commit refactors how django rest will retrieve the initial object so that it's just simply locked from the beginning and we don't have to be so careful.