RHCLOUD-35015: Bring mapping reads within locked transaction to prevent out of order replication (#1180)
* Bring mapping reads within locked transaction to prevent out of order replication

* Bring mapping and initial role read within locked transaction to prevent out of order replication

It turns out that the initial role state was being queried outside of the atomic block and without a row lock and thus subject to a concurrent update causing policy, access, or mappings to not be the current state by the time the outbox event is written, which would create inconsistency. This commit refactors how django rest will retrieve the initial object so that it's just simply locked from the beginning and we don't have to be so careful.

* Omit partial_update from update queryset

partial_update doesn't need the lock or the mapping because it does not update access or policy tables.

* Remove commented parameters

* Remove more commented parameters

* Grammar

* Add note about repeatable read option

* Create mapping if none loaded prior (i.e. for new Roles)

* Create mapping in one SQL statement rather than two

* Consolidate audit in perform_; fix lint and programming errors

* Add back permission check to queryset logic

* Fix whitespace

* Fix typo in assert method

* Fix permissions check in get_queryset

* Add missing fine grained permissions case

* Fix tests that reference renamed method

* Handle when role is updated but binding does not exist

* Go back to original policy delete method (what was I thinking!?)

* Don't bother getting relations if binding does not exist