Ensure paginated links in v2 APIs are supported/consistent with v1
Currently `links` are rendering like so for v2 endpoints:
```
{
  "links": {
    "first": "http://localhost:8000/api/rbac/v2/workspaces/?limit=1&offset=0",
    "last": "http://localhost:8000//api/rbac/v2/workspaces/?limit=1&offset=3",
    "next": "http://localhost:8000//api/rbac/v2/workspaces/?limit=1&offset=2",
    "previous": "http://localhost:8000//api/rbac/v2/workspaces/?limit=1"
  }
}
```

And we need them to be consistent with v1 like:
```
{
  "links": {
    "first": "/api/rbac/v2/workspaces/?limit=1&offset=0",
    "last": "/api/rbac/v2/workspaces/?limit=1&offset=3",
    "next": "/api/rbac/v2/workspaces/?limit=1&offset=2",
    "previous": "/api/rbac/v2/workspaces/?limit=1"
  }
}
```

In order to do this, it requires we change how we patch the `link_rewrite` method.
This could be made simpler if we didn't have the current test cases expecting
links to work for URLs without a `/api/v[0-9]` prefix.

This PR adds an explicit test case for our prefix, and continues backward support
for the existing test cases, though we may want to consider changing these since
I don't believe they're valid use-cases today, in order to simplify the link
logic in the paginator.