Update spec to optionally include ancestry in `GET /workspaces/<uuid>/`
By default, we should just return the workspace detail data on `GET /workspaces/<uuid>/`
but optionally include `/?ancestry=true|false` as a supported query param.

This updates the spec to document the query param, its intent and the expected
response by including ancestry as part of the response as an `anyOf` option.