Ensure we allow `application/problem+json` as a valid `Accept` header value in v2
Ticket: RHCLOUD-35321

Since we are starting to use `application/problem+json` as a response content type
in v2 APIs [1], we need RBAC/Django to accept this as a valid `Accept` header value,
given clients built off the API spec will be sending this.

Currently, if you send a request to `/api/rbac/v2/workspaces/<uuid>/ -XDELETE -H 'Accept: application/problem+json'`
which is what a client built off the spec will do (with the `Accept` header),
you'll get an error from RBAC: "Not acceptable 406".

We _do_ set the correct `content-type` to align with the spec [2], however Django
complains becasue it can't find the `application/problem+json` renderer.

This implements a custom renderer in Django Rest Framework [3] on the v2 views,
specifically workspaces now, specifying it in the `renderer_classes` of the view,
and setting the `media_type` (used to match the `Accept` value(s) to a renderer)
and `format` (response) properties accordingly.

The `DELETE` test was modified prior to this change to add `Accept: application/problem+json`
to confirm it failed (it did), and now succeeds with the changes in place.

[1] [https://github.com/RedHatInsights/insights-rbac/blob/master/docs/source/specs/v2/openapi.v2.yaml]
[2] https://github.com/RedHatInsights/insights-rbac/blob/master/rbac/api/common/exception_handler.py#L93
[3] [https://www.django-rest-framework.org/api-guide/renderers/#custom-renderers]