Disable v2 APIs by default (configurable)
By default, we'll disable v2 APIs, with the ability to enable them through deployment
configuration in order to prevent access prior to full production availability and support.

Locally, you'll need to set your `.env` want to set `V2_APIS_ENABLED=True` in order
to exercise v2 APIs locally.

In stage we'll also want to configure this to set it on via param.