Update remove_role from group algorithm
I'm not sure why I thought it would be the same for system roles and custom roles before. For system roles we can't retrieve all role bindings, that would query across all tenants on the platform. Adding and removing should be basically the same. So, I factored out a method that they could both reuse and just parameterized the few parts that change.