Enable support for read-only mode on APIs
This adds an optional `READ_ONLY_API_MODE` environment variable consumed in settings,
which defaults to `False`. It can be configured via params in the deploy template
to enable/disable with config per environment.

This is used in a new `ReadOnlyApiMiddleware` middleware for all RBAC requests,
which checks the request method and the configuration setting to determine whether
or not to process the request.

In the event of a write method + read-only mode being enabled, we'll return a 405
with a message explaining we're in read-only mode, to be used during migration/maintenance
windows for the service.

Another option would be to use 503, though since we're still accepting GET requests
a 405 with explicit message may be the best approach.