Modify Workspace#parent_id to Workspace#parent_uuid
This updates the `parent_id` column on the `Workspace` table to instead be `parent_uuid`
in order to be more explicit and representative of what it is.

Since we're defining `parent` on the model as a self-referencial FK using the `uuid`
column using `to_field` [1], this will by default add a `parent` association on the model, as well as
a `parent_id` field and column in the database.

In order to modify the column name to `parent_uuid`, we need to use `db_column` [2].
It should be noted that while the _database column name_ will change [3], this _will not_
change the model field in Django, as `ForeignKey.get_attname()` will return `f{field}_id` [4,5].

This complicates things a bit since the model field/attribute names are used in
queries, not the column names. So we can do something where we create an aliased
property `parent_uuid` on the model to return `parent_id`, and a corresponding setter
for `parent_uuid`, but when we filter on queries like `Workspace.objects.filter(parent_uuid=<uuid>)`
this will fail, and we'd still need to use `Workspace.objects.filter(parent_id=<uuid>)`,
which is why this line [6] remains unchanged.

[1] https://docs.djangoproject.com/en/4.2/ref/models/fields/#django.db.models.ForeignKey.to_field
[2] https://docs.djangoproject.com/en/4.2/ref/models/fields/#db-column
[3] https://github.com/django/django/blob/4.2.16/django/db/models/fields/__init__.py#L924
[4] https://github.com/django/django/blob/4.2.16/django/db/models/fields/__init__.py#L89-L90
[5] https://github.com/django/django/blob/4.2.16/django/db/models/fields/__init__.py#L919-L920
[6] https://github.com/RedHatInsights/insights-rbac/blob/592b9c2f2d7e4eeaa277ec5d7c195e6bcb47035d/rbac/management/workspace/serializer.py#L84