Filter default queryset for workspaces (and all v2 viewsets)
In order to ensure we're starting with a default queryset that restricts a v2
model's API response to only records for the requester's tenant, this abstracts
an override of the `get_queryset` method on our existing `BaseV2ViewSet` class
which v2 viewsets inherit from.

We can dynamically set the queryset model by adding a `basename` value [1] which
the queryset uses, filtering to the request tenant.

[1] https://www.django-rest-framework.org/api-guide/routers/