Validate `parent_id` on workspace PATCH
Now that we require `parent_id` on PATCH as well, we need to validate the workspace
is in the correct tenant before attributing the `parent_id`.

This removes an unnecessary check against the `parent_id` since both PUT and PATCH
require it now as part of `validate_workspace`.

We can now reused `validate_workspace` for PUT, PATCH and CREATE.