Revert filter on internal Tenant endpoint
The intent of this endpoint (even if not regularly used now with Gabi) is to have
visibility on all tenants for debugging. In order to preserve existing functionality
we should leave this. The api does respect the default limit of 10, and I've tested
locally against 10 million tenant records that this returns the default in ~ 3 seconds.

We may want to consider whether or not we want to remove this (and other) internal
endpoint altogether now.