Test tenant readiness management
This also changes the behavior of V1 bootstrapping in the cleaner. It will no longer create ready Tenants for Principals that aren't present. Instead, it will not create Tenants at all. This should be left to V2, when we create Tenants unready with bootstrapping.

This may already be occurring in production. If so, it should be okay as the data volume over the cleaner queue is not large.