Lazily mark tenants ready + optimization (#1328)
* Ensure in middleware that tenants are marked ready

This will lazily flag tenants as ready when the first principal for a tenant
makes a request into RBAC. This utlimately will allow us to query on this flag
and inialize tenants with `ready=False` (in a following commit) so that we exclude
those from access cache busting, and seeding notifications.

* Use  filter to only purge cache for tenants with activity.

* Use  filter to only send notifications to tenants with activity.

* Avoid ready-ing Tenants in cleaner and bulk import

* Exclude unready tenants in a few more edge cases

* Also exclude unready tenants in migrator

* Add concurrent index on `Tenant#ready`

This adds an index to the `Tenant` model on the `ready` column, applied concurrently [1]
so that we avoid locking the table while the index is created.

In order to accomplish this, we need to modify the migration which Django creates
by default (which is atomic) with two updates:

1) Make the migration non-atomic [2].
2) Use `AddIndexConcurrently` in the migration explicitly to ensure it's created concurrently

The reverse of this migration does drop the index.

[1] https://docs.djangoproject.com/en/4.2/ref/contrib/postgres/operations/#concurrent-index-operations
[2] https://docs.djangoproject.com/en/4.2/howto/writing-migrations/#non-atomic-migrations

* Fix tenant query and tests in migrator

`Tenant.objects.exclude(tenant_name="public", ready=False)` will only exclude tenants
where `tenant_name="public"` AND `ready=False`, meaning the public tenant will
always be returned.

This updates that query and adds `ready=True` to the tests which create tenants.

* Revert filter on internal Tenant endpoint

The intent of this endpoint (even if not regularly used now with Gabi) is to have
visibility on all tenants for debugging. In order to preserve existing functionality
we should leave this. The api does respect the default limit of 10, and I've tested
locally against 10 million tenant records that this returns the default in ~ 3 seconds.

We may want to consider whether or not we want to remove this (and other) internal
endpoint altogether now.

* Fix additional tenant queries

`Tenant.objects.exclude(tenant_name="public", ready=False)` will only exclude tenants
where `tenant_name="public"` AND `ready=False`, meaning the public tenant will
always be returned.

Instead we should use `Tenant.objects.filter(ready=True).exclude(tenant_name="public")`
to filter to ready tenants and exclude the public tenant.

* Test tenant readiness management

This also changes the behavior of V1 bootstrapping in the cleaner. It will no longer create ready Tenants for Principals that aren't present. Instead, it will not create Tenants at all. This should be left to V2, when we create Tenants unready with bootstrapping.

This may already be occurring in production. If so, it should be okay as the data volume over the cleaner queue is not large.

* Ensure unready tenants are not notified

* Also ensure group seeding does not notify unready tenants

* Simplify asserts

---------

Co-authored-by: Alec Henninger <alechenninger@gmail.com>