Add method to retrieve the public tenant
This adds a method on the `Tenant` model manager, `get_public_tenant()` which can
be called with `Tenant.objects.get_public_tenant()`.

The public tenant record will be added to/retrieved from the cache. A new cache
inheriting from `TenantCache` was added to allow for a custom TTL and key to be
used for the public tenant (open to feedback here, I just didn't want to refactor
too much initially).