Create a custom queryset on the `Workspace` model for built-ins
This adds a custom queryset (though maybe we could also consider a manager?) for
the `Workspace` model, which will help DRY up the queries for built-in types:
(root + default) for a given tenant.

We override/copy the default `objects` manager, and use `.as_manager` [1] to include
the custom queryset method.

```python
built_in_workspaces = Workspace.objects.built_in(tenant=tenant)
```

This can be chained as well, like `built_in.filter(...)` if needed, as it returns
a queryset object.

[1] https://docs.djangoproject.com/en/5.1/topics/db/managers/#creating-a-manager-with-queryset-methods