Add configurable workspace depth limit
This adds `WORKSPACE_HIERARCHY_DEPTH_LIMIT` as a configurable depth limit for the
hierarchy of workspaces. The default is set to 2, meaning off the root, we'll allow
for 2 levels of hierarchy for a tenant: 1 being the default workspace, and 2 being
children of the default.

Open to changing the starting depth value of this if it's confusing.