Update the `ResourceDefinition` serializer to return descendant workspace IDs
We override `to_representation` here to append descendant workspace IDs to the
JSON output conditionally.

If the `ResourceDefinition` being rendered is not for a workspace permission, and
does not have an ID key for the attribute filter, nothing changes.

In the case where it does, we'll collect descendant workspace IDs and append them
to the filter list to be returned to the client.