Add a `FeatureFlags` class to access Unleash feature flags
This adds an evolvable class for accessing feature flags via Unleash. We initialize
the cache/client through config (relying on Clowder when deployed), and allow for it
to be imported and used like:

```
 from feature_flags import FEATURE_FLAGS
 enabled = FEATURE_FLAGS.is_enabled("FeatureName")
```

We override the `is_enabled` method in order to default to `False`, and allow
for a custom fallback function if there's no client initialized.